refactor(core): Plan A — extract shared Gemini client to tools/core/gemini
- tools/core/gemini/client.py  <- tools/summary/batch_runner.py (config import repointed)
- tools/core/gemini/config.py  = the Gemini knobs (model tiering, thinking, caching,
  output-token estimates) + DRYRUN_DIR + LAST_BATCH_FILE (same path values, own PROJECT_ROOT)
- tools/summary/batch_runner.py = full-namespace re-export shim (identity preserved,
  incl. _PENDING_BATCHES) so translator + cli + ~30 monkeypatch sites keep working
- tools/summary/config.py = trailing re-export of the moved knobs (single source = core)
- .importlinter: core.gemini added to the core-is-a-leaf contract

Config-split hotspot (planner-flagged) handled with minimal blast radius:
- LAST_BATCH_FILE/DRYRUN_DIR are now canonical in core.gemini.config (client reads them).
  cli.py is UNCHANGED (reads summary.config, which re-exports core's same objects).
  Test isolation: conftest + the 2 submit tests patch BOTH summary.config AND
  core.gemini.config so the client's reads stay redirected to tmp.
- DEVIATION: did NOT surgically remove the 12 scattered consts from config.py (risky on a
  249-line interleaved file); a trailing re-export import shadows the now-dead local defs
  (identical values). MAX_BATCH_COST_USD/COST_CONFIRM_THRESHOLD_USD stay summary-local
  (the client never reads them; cli's cost gate does).

Gate: lint-imports 1 kept/0 broken; core.gemini.client imports with no tools.summary dep;
shim+config identity verified; targeted 102 passed (batch_runner/cost_opt/translator/cli/e2e);
full suite 425 passed / 2 pre-existing fails.
