+++ .importlinter
@@ -15,6 +15,7 @@ source_modules =
     tools.core.web
     tools.core.content
     tools.core.seo
+    tools.core.gemini
 forbidden_modules =
     tools.summary
     tools.translator
