+++ tools/summary/config.py
@@ -246,3 +246,16 @@ THINKING_BUDGET_TOKENS = 1500
 
 # Webflow API base.
 WEBFLOW_API_BASE = "https://api.webflow.com/v2"
+
+# ---- Gemini config (shared) — re-exported from tools.core.gemini.config (Plan A) ----
+# These knobs are now CANONICAL in tools/core/gemini/config.py (the home the shared
+# Gemini client reads). This trailing import shadows the local definitions above so
+# summary code keeps `config.MODEL_ID` / `config.DRYRUN_DIR` etc. working off the
+# single shared source. (MAX_BATCH_COST_USD / COST_CONFIRM_THRESHOLD_USD stay
+# summary-local — the client never reads them; cli.py's cost gate does.)
+from tools.core.gemini.config import (  # noqa: E402,F401
+    MODEL_ID, MODEL_BLOG, MODEL_BY_CONTENT_TYPE, model_for_content_type,
+    OUTPUT_TOKEN_ESTIMATE, DEFAULT_OUTPUT_TOKEN_ESTIMATE, THINKING_BUDGET_TOKENS,
+    CACHE_TTL_SECONDS, CACHE_MIN_GROUP_SIZE, ENABLE_EXPLICIT_CACHE,
+    DRYRUN_DIR, LAST_BATCH_FILE,
+)
